Rename InnerSidebar to Sidebars. Implement OuterSidebar function. Wrap Switch route component with Sidebar.Pushable and Sidebar.Pusher component.
